+++ Program/Data/Collection/Collection.cs
