+++ Program/Data/DataUnit/DataUnitProp/ArrayProp/DoubleArrayDataUnitProp.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Program
+{
+    public class DoubleArrayDataUnitProp : DataUnitProp
+    {
+        public DoubleArrayDataUnitProp(string name, List<double> value) : base(name, value, DataType.DoubleArray)
+        {
+        }
+    }
+}
