+++ Program/Data/DataUnit/DataUnitProp/ArrayProp/ObjArrayDataUnitProp.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Program
+{
+    public class ObjArrayDataUnitProp : DataUnitProp
+    {
+        public ObjArrayDataUnitProp(string name, List<DataUnit> value) : base(name, value, DataType.ObjArray)
+        {
+        }
+    }
+}
