+++ Program/Data/DataUnit/DataUnitProp/ArrayProp/RefArrayDataUnitProp.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Program
+{
+    public class RefArrayDataUnitProp : DataUnitProp
+    {
+        public RefArrayDataUnitProp(string name, List<DataUnitReference> value) : base(name, value, DataType.RefArray)
+        {
+        }
+    }
+}
