+++ Program/Data/DataUnit/DataUnitProp/ArrayProp/StringArrayDataUnitProp.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Program
+{
+    public class StringArrayDataUnitProp : DataUnitProp
+    {
+        public StringArrayDataUnitProp(string name, List<string> value) : base(name, value, DataType.StringArray)
+        {
+        }
+    }
+}
