+++ Program/Data/DataUnit/DataUnitProp/SingleProp/RefDataUnitProp.cs
@@ -1,0 +1,9 @@
+﻿namespace Program
+{
+    public class RefDataUnitProp : DataUnitProp
+    {
+        public RefDataUnitProp(string name, DataUnitReference value) : base(name, value, DataType.Reference)
+        {
+        }
+    }
+}
