--- Program/Data/DataUnitProp/SingleProp/DoubleDataUnitProp.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace Program
-{
-    public class DoubleDataUnitProp : DataUnitProp
-    {
-        public DoubleDataUnitProp(string name, double value) : base(name, value, DataType.Double)
-        {
-        }
-        protected override List<byte> GetValueBytes()
-        {
-            return new() {Convert.ToByte(Value)};
-        }
-    }
-}
