--- Program/Data/DataUnitProp/SingleProp/IntDataUnitProp.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace Program
-{
-    public class IntDataUnitProp : DataUnitProp
-    {
-        public IntDataUnitProp(string name, int value) : base(name, value, DataType.Integer)
-        {
-        }
-        protected override List<byte> GetValueBytes()
-        {
-            return new() {Convert.ToByte(Value)};
-        }
-    }
-}
