--- Program/Data/DataUnitProp/SingleProp/RefDataUnitProp.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Program
-{
-    public class RefDataUnitProp : DataUnitProp
-    {
-        public RefDataUnitProp(string name, DataUnitReference value) : base(name, value, DataType.Reference)
-        {
-        }
-        protected override List<byte> GetValueBytes()
-        {
-            var val = (DataUnitReference) Value;
-            return val.GetBytes();
-        }
-    }
-}
