+++ Program/Program.csproj
@@ -6,4 +6,8 @@
         <UseWPF>true</UseWPF>
     </PropertyGroup>
 
+    <ItemGroup>
+      <Content Include="obj\Debug\net5.0-windows\MainWindow.baml" />
+    </ItemGroup>
+
 </Project>
